Move validator from utils to api and fix the regex pattern
