+++ internal/controlplane/server.go
@@ -39,6 +39,7 @@ import (
 	"google.golang.org/grpc/reflection"
 	"google.golang.org/grpc/status"
 
+	"github.com/mindersec/minder/internal/api"
 	"github.com/mindersec/minder/internal/assets"
 	"github.com/mindersec/minder/internal/auth"
 	"github.com/mindersec/minder/internal/auth/jwt"
@@ -234,7 +235,7 @@ func (s *Server) StartGRPCServer(ctx context.Context) error {
 		return fmt.Errorf("failed to listen: %v", err)
 	}
 
-	validator, err := util.NewValidator()
+	validator, err := api.NewValidator()
 	if err != nil {
 		return fmt.Errorf("failed to create validator: %w", err)
 	}
@@ -243,7 +244,7 @@ func (s *Server) StartGRPCServer(ctx context.Context) error {
 	interceptors := []grpc.UnaryServerInterceptor{
 		// TODO: this has no test coverage!
 		util.SanitizingInterceptor(),
-		util.ProtoValidationInterceptor(validator),
+		api.ProtoValidationInterceptor(validator),
 		// This adds `Grpc-Metadata-Request-Id` to the
 		// response.
 		logger.RequestIDInterceptor("request-id"),
