+++ proto/minder/v1/minder.proto
@@ -2313,9 +2313,9 @@ message RoleAssignment {
     // subject is the subject to which the role is assigned.
     string subject = 2;
     // display_name is the display name of the subject.
-    string display_name = 5 [(buf.validate.field).string.pattern = "^(?!.*<\\/?[a-zA-Z][^>]*>).*$"];
+    string display_name = 5 [(buf.validate.field).string.pattern = "[A-Za-z][- [:word:]\\[\\]\\(\\)]*"];
     // project is the project in which the role is assigned.
-    optional string project = 4 [(buf.validate.field).string.pattern = "^(?!.*<\\/?[a-zA-Z][^>]*>).*$"];
+    optional string project = 4 [(buf.validate.field).string.pattern = "[A-Za-z][- [:word:]\\[\\]\\(\\)]*"];
     // email is the email address of the subject used for invitations.
     string email = 6 [(buf.validate.field) = {string: {email: true}}];
     // first_name is the first name of the subject.
